+++ src/ffmpegio/caps.py
@@ -1,7 +1,9 @@
 # TODO add function to guess media type given extension
 
 import re, fractions, subprocess as sp
-from .path import get_ffmpeg
+
+from .path import where
+from .utils.error import FFmpegError
 
 # fmt:off
 __all__ = ["options", "filters", "codecs", "coders", "formats", "devices",
@@ -28,7 +30,15 @@ _cache = dict()
 
 
 def ffmpeg(gopts):
-    return sp.run([get_ffmpeg(), '-hide_banner', *gopts], stdout=sp.PIPE, encoding="utf-8").stdout
+
+    out = sp.run(
+        [where(), "-hide_banner", *gopts], stdout=sp.PIPE, encoding="utf-8"
+    )
+
+    if out.returncode or out.stdout.count("\n") == 1:
+        raise FFmpegError(out.stdout)
+
+    return out.stdout
 
 
 def _(cap):